Kill FPassCE and FPassCW
Summary:
Stop reporting warnings and errors for some types of expressions when they
produce a cell and are being passed to a function that expects that argument
to be a ref.

This gets rid of usage of the mustBeRef() logic by FPass* opcodes which
required access to the pre-live ActRec which we are trying to get rid of. After
this diff, FPass* opcodes need ActRec only for the byRef() logic.

These warnings and errors are not very useful for a number of reasons:

- this logic was very arbitrary:
-- it was already disabled for 4 random helpers (extract, current, key, array_multisort)
-- many types of expressions that produce a cell were whitelisted and did not produce any warnings or errors
- removal of this logic won't break any existing code; it will just make HHVM more permissive, potentially allowing more nonsensical code to run
- HHVM supports call-side `&` annotations which should be used instead to avoid nonsensical code
- support for references is deprecated in favor of inout

Reviewed By: markw65

Differential Revision: D8046132

fbshipit-source-id: fde6f9ebc88f36727bc4a29fe87110e9f24f807f