Rename Flavor::CUV to Flavor::CU
Summary: FlavorDesc values in hhbc.h all end in V, but the corresponding values of Flavor in hhbbc/misc.h don't have the V. This pattern did not hold for the CUV FlavorDesc: the corresponding Flavor was also called CUV and this was confusing. Rename it and make sure it's in the same order WRT other flavors in Flavor as it is in FlavorDesc.

Reviewed By: markw65

Differential Revision: D8246394

fbshipit-source-id: 5bcbd80c1a0f61687b7a3deede4a5372b0070ff6