Enable WarnOnCallByRefAnnotationMismatch by default
Summary:
Enable warnings for mismatches between callsite `&` annotations and reffiness
of target functions in `<?hh` files. Update all affected unit tests either by
adding missing annotations (preferred), changing expected test output (when the
test tries to test for this behavior, or fixing the test is non-trivial),
or disabling the option for a specific test (when the test inherently cannot
work in this way).

Reviewed By: markw65

Differential Revision: D8246339

fbshipit-source-id: d66c97e2465502ffb3db8375e5445cc019a46f4c