LSP supports findReferences
Summary:
This adds support for textDocument/references

Note: flow currently has two forms of find-references - single-hop and multi-hop. We have to decide which one of these the default VSCode "find-references" will jump to. At the moment I've made it default to multi-hop... that's because of my hunch that the way folks write Flow code is that multi-hop is more valuable. But it could have gone the other way.

Separately, it would be nice to have a FB-specific customization of textDocument/references so that Nuclide users can explicitly chose whether they want single-hop or multi-hop. I'll leave that to a future diff. (it requires adding the functionality into AtomLanguageServiceConfig and LanguageService in Nuclide, and after than inventing a custom addition to the LSP protocol in both Nuclide and Hack+Flow).

Reviewed By: nmote

Differential Revision: D8175729

fbshipit-source-id: b39de486d6e8a9f74a1b2e8cf8306d8c022d2daf