LSP supports documentSymbol
Summary:
This diff adds support for textDocument/documentSymbol

In VSCode, documentSymbol appears when you type Cmd+P; it provides a flat list of all useful named symbols in the document so you can review/filter the list and jump to one easily.

In Nuclide, documentSymbol is what we use to reconstruct Outline (with the assumption that if symbol A's range is contained within symbol B's range then symbol A is a child of symbol B), and also the Breadcrumbs bar.

(VSCode are in the process of introducing a new LSP specifically aimed at returning hierarchical lists of symbols, but they're not there yet. And in any case we'll always have to support documentSymbols).

Previously in Nuclide, OutlineView was obtained by invoking `flow ast`. This was performed by astCommand -- it didn't call out to the flow server. It turned the entire AST into a JSON tree. Nuclide's Flow module obtained the JSON tree and had a bunch of code to extract out the symbols that should be displayed. Nuclide has the potential to recompute the outline tree on every keystroke, but this had to be turned off because it was too slow.

The code I've added is largely equivalent to what Nuclide was doing. The difference is I'm avoiding the JSON intermediary, and I'm able to walk a much smaller portion of the tree.  Also, since this is LSP, we're serving the request over an existing persistent connection rather than burning 25ms just to spin up the flow executable to run `flow ast`. I hope this way it can be faster.

Note: the ontology of SymbolInformation kinds is proscribed by LSP. We can't add to it with Flow-specific things (indeed, the various editors that speak LSP wouldn't even have icons if we did add anything).

Note: there's no mandate about what things must be shown in Outline View. I chose the things I did with the goal that what we see in the Outline view should be user-relevant, and by comparison with the existing Nuclide codebase to see which ones it was showing. Maybe I've missed some for the best user experience, or maybe I've gotten too many. We'll see how it pans out in practice with the Outline View and with Breadcrumbs.

Reviewed By: nmote

Differential Revision: D8177564

fbshipit-source-id: ae07cf92fe721426cc3c5563972e054d79d83746