Optimize 'or $0x0 ...' instructions in vasm simplifier.
Summary:
In some generated code instructions like 'or $0x0 %r13' can be optimized to
nop and copy instruction when the flags are not used. Also when two registers
both two constant values, the result can be evaluated at compile time. A new
simplifier rule is added to handle these cases in orq instructions.

Reviewed By: swtaarrs

Differential Revision: D8211889

fbshipit-source-id: f0bd2b9f989e0e0095e2615c8b67ce429ad22a0b