Disallow accessing class consts and static methods and variables on an unconstrained generic type parameter
Summary:
The following example is currently accepted by the type-checker:
```
<?hh // strict

function test<T>(classname<T> $c): void {
  $c::doesNotExist();
}
```
where `doesNotExist` stands for an arbitrary method name.  I.e., we don't check if `$c` is the name of a class that defines a static method `doesNotExist()`.  Ditto for class consts and static variables.  The reason for this type hole is that when the generic type parameter `T` is unconstrained (or has more than one upper bound) we effectively give `$c` the "any" type `_`.  Let us instead do something similar to how object property accesses are handled: iterate over the list of concrete supertypes and try to lookup the method (const, variable) in those supertypes.  If there is only one successful result, we are good.  If there are more than one, the access is ambiguous.  If there are none, we are trying access a property that may not exist.

I mimicked the error messages after the `Obj_get` case but I'm not 100% happy with them and welcome suggestions for improvement.

Reviewed By: CatherineGasnier

Differential Revision: D8203378

fbshipit-source-id: 813e3d67e963693de99851b79fb45b7a0605ab1e