Prevent job cancellations in the middle of declaration
Summary:
Job cancellations work by worker periodically calling `check_should_exit()` function. This adds a facility to disable it in parts of code where we don't want the worker to exit. I then use it to make sure that we don't interrupt class declarations, which write multiple things to shared memory (i.e. declaring a class will write to `Decl_heap.Classes`, `Decl_heap.Constructors`, `Decl_heap.Methods`, etc...). We want them to happen "atomically" and avoid the situation where only half of them end up being written.

I initially didn't implement it, because I though that it's enough to maintain an invariant "if the class is in memory, then it's elements are there too" (which is maintained, because we write elements before the class).

But the test failures in D8129532 (unrelated to this project, but dealing with similar issues of changing what we have/don't have in memory) made me realize that it must be "if and only if" - if we write the elements, but cancel before writing the class, the elements will be abandoned in memory and possibly interfere with subsequent incremental mode operation (if they change, we'll try to write the new version in a place that is already occupied, which is a no-op).

Reviewed By: ljw1004

Differential Revision: D8250495

fbshipit-source-id: 99903e2fa4d74ff715d503579e8e7e7c48412233