Fix propagation of known refcounts through phi nodes
Summary:
I attempted this in D7601541 a few weeks ago, but got it completely
wrong.

That version attempted to pull refcounts from preceding blocks into
the DefLabel block, but failed because it was inadvertently looking at
the known state of the input *in the current block*, not at the end of
the preceeding block. Fortunately, the upshot of that is that it very
rarely did anything, and in the cases where it did, it was
conservative. It turns out we don't store block exit states anywhere,
so this reverses it to propagate the state forward when it sees a
phi-jmp.

Reviewed By: swtaarrs

Differential Revision: D8097027

fbshipit-source-id: 150cc0b560263aa6fc54f69a62eaabd3dcde27e1