HHVM Debugger: Fix bug in request attach flow and increase logging
Summary:
This diff fixes a bug that was causing the debugger not to attach to certain requests.  When the debugger client connects, the debugger attaches the debugger hook to any request threads that already exist, and then sets a flag in an internal (to the debugger) data structure that keeps track of request info, keyed off ThreadInfo* for the request thread.

There was a bug in this logic that caused us to attach the hook, but then lose the hook in certain cases. If a worker thread exists, but is not running any request (so it's just parked in the HHVM thread pool doing nothing) when the debugger client attaches, we attach the hook and set a flag not to attach it again.  Then, when that thread picks up a request and BEFORE it calls requestInit() in the extension framework, HHVM calls reset() on the threads RequestInjectionData, which clears the debugger hook flag.  In requestInit(), the debugger doesn't re-attach the hook (because it thinks it's already attached to it) and the request proceeds without the hook flag set. As a result, the VM doesn't invoke debugger callbacks for the request, causing us to miss breakpoints. This repros only once per background thread that existed when the client attached but started a request afterwards.  After the request ends and the thread is re-used for the next request, we'll attach because the debugger's request shutdown handler clears the flag.

This diff simplifies the logic by removing hookAttached from debugger flags, and looks at the thread's request injection data directly, as the single and authoratative source of whether or not the debugger hook is installed on the thread.

Also adding more logging so these sorts of issues are easier to diagnose in the future.

Reviewed By: velocityboy

Differential Revision: D8194469

fbshipit-source-id: 0ec22d771532cad5801e9dce7ad741612d6e4fea