Fix potential use-after-free in gen-type-scanners.cpp
Summary:
We were moving a member of Layout out of it and overwriting the same
layout with that member, all in one expression. This has indeterminate
evaluation order, leading to possible use-after-frees. Fix it by moving the
member out into a temporary, and then moving that temporary into the Layout.

Reviewed By: edwinsmith

Differential Revision: D8267088

fbshipit-source-id: 1b258093d191bea5092fbc55b5f1df88bea55af4