Fix hack coverage
Summary:
There seems to be a regression, maybe from D7845027, which prevents Hack from reporting coverage levels to Nuclide.

This diff restores the type-coverage functionality.

I didn't try to recreate the prior algorithm which merged uncovered-ranges together if they overlapped. Instead I switched to what I think is a better algorithm where, if an uncovered range contains another uncovered range, then only report the smaller of the two. For instance,
```
$user = $vc->getUserID()
// if $vc is uncovered then hh_server provides 3 uncovered regions
// 1. $vc is uncovered
// 2. $vc->getUserID() is uncovered
// 3. $user = $vc->getUserID() is uncovered
```
I think that it's more actionable to the user if we squiggle the `$vc` rather than the `$user = $vc->getUserID()`.

Reviewed By: pittsw

Differential Revision: D8199212

fbshipit-source-id: 0bb0fe3737a39d6f5c83548e0a2cb01660adf763