Add `--vendor` option to `hphp/test/run`
Summary:
We use hphp/test/run to test the docs.hhvm.com examples, including in `--typechecker` mode (which is not used elsewhere).

We want to start using the HSL in these examples, which means we need to pull in extra code. Copying the HSL in is prohibitively slow - so:
- for runtime, pull in `$VENDOR/hh_autoload.php` to make the definitions available at runtime without copying in or needing to re-parse
- tell the typechecker to also pull in definitions from the vendor path

Also, decrease the thread count for `--typechecker` mode as running it is basically burning my machine.

Reviewed By: alexeyt

Differential Revision: D8170605