Move some subtyping checks involving an optional supertype to simplify_subtype
Summary: The reduction of `?ty_sub <: ?ty_super` to `ty_sub <: ?ty_super` already happens in `simplify_subtype` (so in particular it can be removed from `sub_type_unwrapped_helper`).  This diff moves the checks in which the supertype is optional and the subtype doesn't contain `null` to `simplify_subtype`.  This reveals a bit of sloppiness in the existing code (see inline comments below).

Reviewed By: andrewjkennedy

Differential Revision: D8235817

fbshipit-source-id: 34819dfe2fb39b69343e22b1109915f0851f1812