Allow comparisons between objects of type DateTimeImmutable
Summary: We allow comparisons between values that are both subtypes of num, or string, or `DateTime`, but apparently there is one more case we need to support: `DateTimeImmutable`, an immutable wrapper around `DateTime`.  I chose to allow mixed comparisons too, i.e., when one object is of type `DateTime` and the other is of type `DateTimeImmutable`.

Reviewed By: andrewjkennedy

Differential Revision: D8254665

fbshipit-source-id: 142a66927bca4fede578957e436ae9e2f0e16476