Remove assert that is causing "safe" crashes
Summary:
D7859834 changed the way we emit code during retranslate all.  It started grabbing the code block for a func prologue twice.  Once when emitting the prologue, and once when publishing it.  Unfortunately it could return different code blocks in the event that the Hot section fills up.
This causes us to get the main code block while publishing a prologue that was emitted into the hot section.  Since gdbRecord only cares if the code block is cold, this should not impact its logic.

Reviewed By: markw65

Differential Revision: D8235961

fbshipit-source-id: 0dc45e809f6142ff0d1eee3993728566d1700d83