Hack returns current or previous position based on whether it points to a trigger character
jSummary:
From the band-aid regression in D8062267, hack used to rely on whether a field specified a request was from nuclide or not to decide whether to return the current or previous position. This now checks whether the current position itself points to a trigger character before returning it, otherwise giving the previous position.

Nuclide has also been modified in D8227027 to give the correct position according to LSP. This edit accounts for this while preventing any current versions of Nuclide from breaking.

After the modification in Nuclide has deployed, this check will be removed and only the previous position will be given.

Note that in this case, trigger characters are denoted by semicolons ";" and closing curly braces "}". Each of these trigger characters offer a variety of formatting features. For ex:

- Typing two statements that end with a ";" from left to right on the same line should cause the rightmost statement to automatically shift to the next line.
             foo(); foo() //Before typing the finishing ';'

             foo();  //After typing
             foo();

- An If statement with improper spacing will be correctly spaced after giving the finishing "}". The "}" will also automatically shift to the next line.
            if(true){ //Before closing curly braces

            if (true) { //After typing the closing curly braces
            }

The test plan refers to a specific instance where these formatting features break due to the difference in the position being given by VSCode (i.e. the position expected by LSP) and the position given by Nuclide.

Reviewed By: arxanas

Differential Revision: D8227541