Use fileinfo object to grab list of files instead of indexer in ServerCoverageMetric
Summary:
We were using the indexer here, which actually makes it impossible for us to write tests for coverage, and also is likely much slower than using the fileInfo, which is guaranteed to only have Hack files.

Note that ServerCoverageMetric is not the same as ServerColorFile, which we use for type coverage in Nuclide.

Reviewed By: dlreeves

Differential Revision: D8292345

fbshipit-source-id: 73c7e56cb20419e97955c1d96f635ed6ec433bf7