Eliminate static locals in generated XHP methods
Summary: Methods xhpAttributeDeclaration, xhpCategoryDeclaration, and xhpChildrenDeclaration (prefixed by two underscores) are generated by the parser while handling XHP classes. This is done in both HHVM's parser and HackC. All these methods use a static local as a cache. This change eliminates these static locals in both HHVM's parser and HackC. xhpAttributeDeclaration is modified to use a private static property instead. The other two functions are updated so that they return a value directly, without a cache. This is efficient because both methods return a static array.

Reviewed By: paulbiss

Differential Revision: D8219743

fbshipit-source-id: b8d18e3f8d348f6745341c5b3b9788f6876df9c4