Expose in_loop in Tast_env
Summary:
Because of T25871119, some linters need to be conservative when inside loop bodies. Tautology checks, for example, might warn about a `=== null` check always being false, when it's actually true on a different loop iteration.

In order to allow these linters to act more conservatively in loop bodies, this change sets and exposes the Typing_env.in_loop field in Tast_visitor and Tast_env.

Reviewed By: pittsw

Differential Revision: D8319243

fbshipit-source-id: 1cfd4bbbcf2125bc633bd8c854cb231beccfda4d