HHVM Debugger: Add debugger hook callout for function intercepts
Summary: This diff adds a mechanism for an attached debugger to be informed when an intercept handler is registered or unregistered for a function. The debugger can use this information to inform the user when a breakpoint will not be reachable due to an intercept.  This is going to help us out in cases where testing or mocking frameworks prevent breakpoints from being hit (and code from being run) leading to user confusion.

Reviewed By: velocityboy

Differential Revision: D8230602

fbshipit-source-id: fc3201043cca5668260ea98b127f9af979432f62