HHVM Debugger: Add option to warn if breakpoint is in intercepted function
Summary:
This diff adds an option the debugger client can send as part of launch/attach arguments that will opt-in to warning the user any time a breakpoint resolves to a routine in a request that looks like it has an intercept handler registered.  Intercept handlers can be registered and are used for things like test frameworks that mock out various classes, or stub out functions. This can lead to user confusion. In some cases, the user sets a breakpoint in an intercepted routine. It resolves in the real compilation unit, but since program execution will never enter that unit, it never hits - resulting in it appearing as if the debugger does not work.

Adding this messaging will help users understand what is going on and allow them to more easily unblock themselves

Reviewed By: velocityboy

Differential Revision: D8230601

fbshipit-source-id: 16416d5a0fde230f58531b6522ec2cd5a13b0a77