HHVM Debugger: Show console output explaining when breakpoints are moved by calibration
Summary:
We calibrate breakpoints if the requested line is not executable, or the source table indicates the nearest executable line is different. In the case of multi-line expressions, the start and end lines differ, and the debugger client shows the breakpoint on the end line.  These situations can be confusing to users who don't understand how the VM actually works, and make it seem like the debugger put the bp in the wrong place.

This diff adds output to tell the user that this is intentional and try to explain what is going on.

Reviewed By: velocityboy

Differential Revision: D8242349

fbshipit-source-id: 5b4a24c5e3cc5dd4704b7990f9d559256e4463d0