HHVM Debugger: Add support for warning users when a bp resolves to a memoized function
Summary:
If a breakpoint is set in a function annotated with Memoize, it can lead to user confusion because the function is not actually executed again by HHVM, even if the user invokes it again, when the same parameters are passed to the routine.

This diff adds a console warning that tells the user the function is memoized, and provides a link to the HHVM documentation for more info.

Reviewed By: velocityboy

Differential Revision: D8261897

fbshipit-source-id: 0054a92d6a3645e919ad0e63655c42f4122cb9a2