Allow limiting the number of elements in a syntax list in `ListPattern`
Summary: If we can't limit the number of elements to match, then we don't have any way of matching only syntax lists with a fixed number of elements: we would always match any syntax lists that happen to match in the first few elements and then have extra elements. This commit adds a field that allows us to limit the maximum length of the list to match.

Reviewed By: pittsw

Differential Revision: D8324411

fbshipit-source-id: 8a9dc8a43b89f688e3cc90a30afcd33f75fda2aa