Remove redundant logging about extern compiler fatals.
Summary: We cache units that result in the extern compiler returning an error politely (as opposed to completely falling over). It is currently not possible to unit test these scenarios, because on the first run we produce both a log about the fatal while compiling, and then the fatal from the unit, whereas on the second run we only get the fatal from the cached unit (i.e. there is no single test expectation that always matches). The logging is redundant, so remove it to enable writing tests that involve extern compiler fatals.

Reviewed By: oulgen

Differential Revision: D8326596

fbshipit-source-id: b479291928eeae3d7183b276dadd70afc783e8d2