Create naming error for attribute classes conflicting with regular classes
Summary: As part of the migration to typed user attributes, we need to prevent the creation of typed attribute that have the same name as regular Hack classes, even though they have the temporary `__Attribute__` prefix. This diff bans classes of the form `X` and `__Attribute__X` from coexisting. This error will be deprecated once namespacing is implemented for attributes in the FFP and codegen, and the prefix is dropped

Reviewed By: jamesjwu

Differential Revision: D8328167

fbshipit-source-id: 0009b7c85d279dc5e4b92ecf0957f33f25805fd2