Mangle HackCompilerVerboseErrors into unit md5.
Summary: The HackCompilerVerboseErrors flag controls the verbosity of fatals in fatal units created due to failures to compile using hackc. These fatal units may be cached (depending on the type of error). As a result, we need to mangle this flag into the unit md5 or it will appear to have no effect when we have a cache hit for a unit.

Reviewed By: paulbiss

Differential Revision: D8339720

fbshipit-source-id: 114ac6666d58f09ad389ed75edf92e32202e09db