Make an hphpc-only test also pass under hackc.
Summary:
Fix the syntax error in hphp/test/slow/parser/memory_exhaust.php.
Fix the assembler so that it wraps any exceptions that aren't AssemblerError in an AssemblerError.
Delete opts files that make it only run under hphpc, replace them with ones that turn off verbose errors in hackc (to suppress printing all the php and hhas in the fatal that hackc produces), and convert expectations to .expectregex so that we can accept the output produced with either compiler.
Add a .nodumphhas file so that we don't attempt to compare generated hhas (the strings in the fatals differ).

Reviewed By: oulgen

Differential Revision: D8306189

fbshipit-source-id: ffafab335138d1a4b8e030bcc0b0e6d3f3c8f6df