Use the same code view for both emitting a prologue and publishing its metadata
Summary:
We were obtaining the code view again for publishing the metadata, but the view
could be different from when we emitted the prologue in case code.hot filled up
in between.  We now save the code view in PrologueMetaInfo when emitting the
prologue, and then just look it up from then for publishing the prologue
metadata.

Reviewed By: markw65

Differential Revision: D8343172

fbshipit-source-id: e817ea5b99c729508a87952800da0cc0b0ad2e31