Add cases for tuples to simplify_subtype
Summary: What it says on the tin.  There is a subtlety compared to the previous diff in the stack: tuples of different lengths cannot be in the subtyping relationship, so this can be decided without resorting to unification, but currently we do just that to report a better error (than "This is a tuple... It is incompatible with a tuple...").  To remedy that, I tried to change the way the tuple type is displayed (so that the arity/length is part of the textual representation).  If this idea is met with objection, we can simply remove the case where both sub- and supertype are tuples from `simplify_subtype` and defer to unification as we currently do.

Reviewed By: andrewjkennedy

Differential Revision: D8333098