Don't assert that a TBottom is defined at its use site
Summary:
Both the def and the use are unreachable (note that the instruction that produces the def may not itself be unreachable - it might eg be a CheckType; but in that case, the CheckType can never pass, and the def never happens), so its not clear which bits of the control flow graph might have been removed, or incorrectly added.

Normally these would be cleaned up by const_prop() anyway, but it looks like some of them slip through (or we check the ir before const_prop runs).

Reviewed By: swtaarrs

Differential Revision: D8338333

fbshipit-source-id: ec705e28e8d8ee142767a0b83db3c86ac309415d