HHVM Debugger: Remove extra stopped event when a breakpoint is hit
Summary:
Currently we're sending two stopped events for a thread hitting a breakpoint - one when the debugger transitions its state of the program from running to paused, and a second for the thread that hit the breakpoint when it enters its command queue. This is techincally OK from the protocol documentation perspective, but in practice it's confusing the UI in Nuclide.

The first stop event is unnecessary and can be removed to simplify this.

Reviewed By: mostafaeweda

Differential Revision: D8283212

fbshipit-source-id: d09967deb0e743328e3d71d9a5f29a75998f0336