LSP live syntax errors
Summary:
Something that's been bugging me is when I make a silly syntax error in a file who's fanout is costly to typecheck, and then do Cmd+S, then it can take 60 seconds before I get told about about the syntax error, and after fixing it it's a further 60 seconds before the syntax error report goes away. It's hard to detect syntax errors without the help of Prettier so I do Cmd+Shift+C to format, but prettier is unable to work when there are syntax errors (and it pops up an ugly red dialog box).

What would make my workflow faster is if I get errors reported as-I-type.

Nuclide/hack and VSCode/typescript and C# and LSP/ocamlmerlin all report errors live as-you-type without the need to save.

The way Hack works is interesting because it's the only one of those with a global singleton server. It does this by having hh_server consider your editor open buffers as the "source of truth" for their contents while it performs its whole-program analysis. Part of Hack's architecture to achieve this involved work to get around the thorny problem that when you edit a file it might cause errors in other files. One problem with the overall user experience was that folks got confused when they did hack commands at the command-line - they didn't understand why the reported errors reflected unsaved changes in an editor that they'd forgotten about. We added a message so when you type 'hh' at the command-line then it prints the message "These errors reflect the state of unsaved files in your editor". Oh, and this setup also means that you can't have two editors open, editing the same hack file, because the source of truth can only come from one of them. (These problems are Hack-specific because it deals with a single global server; they don't affect other languages+IDEs).

The way Visual Studio works is interesting. It has "red squiggles" for live errors that are produced as-you-type error analysis on the current file. And it has "blue squiggles" for errors that are produced through a global typecheck, for all files in the project. Examples of the latter include errors like "this program lacks a main() entry point" which obviously can't be ascribed to a single file.

-----

What I've done is an efficient minimal solution in Flow that does as best we can without running into the Hack user-experience problems: the only live errors it reports are syntax errors for the currently open files. After all, there's no ambiguity about which file to report them on. And parsing is a fast operation that doesn't even require the flow server to be running. And parsing is needed in any case to display the Outline tree and Breadcrumbs feature.

In future I'd like to add different-colored squiggles to Nuclide for live vs built errors, like VS. But that'll be work for the future, and would require additions to LSP.

I'd also be interested in having "as-you-type" error reporting deferred until you complete the line you're working on (by cursoring off, or pressing enter, or clicking elsewhere). This is how Visual Basic works, and would make users happier, but I've not seen any other editors that work like this. It would require changes to Nuclide and to LSP.

I figure that some folks will like live errors, and other folks won't. So it's protected by a user-configurable option. I'll have to decide, based on early-adopter-feedback, whether it should default to be on (so as to get faster user workflows and better parity with VSCode) or default off. But I'll leave it on by default for the early adopters as a forcing function to get feedback from them.

Reviewed By: nmote

Differential Revision: D8273118

fbshipit-source-id: 0303cb6e43f9f49c261d8ad3de15fdf233636fda