Fix the ordering of m_sinitVec and m_pinitVec
Summary:
cls->get86sinit() and cls->get86pinit() return the last elements of m_sinitVec and m_pinitVec respectively.
The assumption is that these will be the s86sinit and s86pinit (respectively) of cls, when these exist.
Ensure that this is true, by pushing the trait initializers before cls's init functions.

Reviewed By: markw65

Differential Revision: D8323342

fbshipit-source-id: 1427047460af725a612a3972c4c255711b026c9a