Disallow `new Foo()->bar()` syntax with FFP
Summary:
As in title.
This syntax has been causing troubles for quite a while due to different behaviours in parsers.
HHVM parser does not parse it.
Legacy Hack parser/FFP parse it as if it is`(new Foo())->bar()`.

We want to make sure that `new Foo()->bar()` is not allowed, users are expected to add parentheses around the constructor call.

Implementation:
Look-ahead after constructor call for `->` token, and report error if so.

Reviewed By: nbenton

Differential Revision: D8378252

fbshipit-source-id: ad44a029161631890f8821454f7bfd5681a427c9