Enable looping with foreach over containers of unresolved types
Summary:
Surprisingly, we don't support `foreach` over containers of unresolved types (unless the keys and/or values happen to have the same type); see the new test cases below, which fail without changes from this diff.  The problem is that when we type-check e.g.
```
foreach ($vs as $v) {...}
```
we introduce a fresh type variable, not fresh //unresolved// type variable for `$v`, and impose the condition that the types of all elements in `$vs` must be subtypes of that type variable (the type of `$v`).  If the type is not unresolved, it cannot grow: the subtyping check will fail as soon as we discover values of incompatible types.  The fix is to use unresolved type variables instead.

Reviewed By: andrewjkennedy

Differential Revision: D8379574

fbshipit-source-id: a28c5e9f15eb953743ec669713835b74c4d14147