Fix typing of Shapes::idx() with a default value
Summary: When type-checking a function call `Shapes::idx($shape, $field, $default)`, we introduce a fresh unresolved variable `Tr` for the result and require that type of `$shape` be a subtype of `shape(?$field => Tr, ...)` (which is a concise way of expressing the requirement that the type of `$field` in `$shape` is a subtype of `Tr`), and that `Tr` be a subtype of a singleton unresolved containing the type of `$default`.  The latter is odd and is actually backwards: the result type variable should be a supertype of both the type of the field in the shape and the type of the default value.  The current implementation works only because of how we implement unresolved types.  In particular, by wrapping the type of `$default` into an unresolved we ensure that the subtyping check of `Tr` against the type of `$default` will actually unify two unresolved types, which will cause `Tr` to grow, as desired.  This is morally wrong though, and this diff fixes it.

Reviewed By: andrewjkennedy

Differential Revision: D8379501

fbshipit-source-id: d2eeafa323aea4aa89582ee31d88916f1b444bee