Return options from Tast_env.get_self_id and Tast_env.get_self
Summary:
Currently, when not in a class definition, Tast_env.get_self_id returns "", and Tast_env.get_self returns a Tany type. This behavior is inherited from Typing_env.get_self_id and Typing_env.get_self, but seems somewhat hazardous. A consumer of get_self_id, for instance, might store that ID somewhere, and end up needing to debug why sometimes the stored IDs are empty strings.

Since it is a logic error to make use of this empty string or Tany type, this change has Tast_env throw exceptions instead. Versions of these functions which return an option instead of throwing are also provided (which is convenient for Lint_utils, which needed such a function).

Reviewed By: pittsw

Differential Revision: D8269844

fbshipit-source-id: 29156a8973620bfc5e38ce8cc9db87439857847e