Remove a redundant unresolved
Summary: Knock off one redundant instance of `Typing_utils.unresolved`.  Here is what we currently do.  We have a list of types `t1`, ..., `tn`, and we want to compute their common supertype (lub?), which we do as follows: we introduce a fresh unresolved type variable `t` for the supertype, wrap each `ti` into an unresolved bound to a fresh variable (if the type is already an unresolved we just rebind it to a fresh variable), and impose the subtyping constraints `t1 <: t`, ..., `tn <: t`, which causes `t` to grow.  However, the wrapping of unresolved around `ti`s seems redundant.  Theoretically, we only care that `t` satisfy all the inequalities `t1 <: t`, ..., `tn <: t`, and singleton unresolved's shouldn't affect that.  Our concrete implementation also makes unresolved's unnecessary: if `t` is a resolved type, then checking `(t') <: t` is equivalent to checking `t' <: t`; and if `t` is an unresolved, then the case `t' <: t` is translated by our current implementation of subtyping into `(t') <: t`.  The only non-trivial effect of using `Typing_utils.unresolved` is rebinding a type to a fresh type variable, so that the original type is not affected by unifications.  Fortunately, all instances of `compute_exprs_and_supertype` are used with functions that internally "unbind" their types, which has a similar effect to rebinding to a fresh type variable.

Reviewed By: andrewjkennedy

Differential Revision: D8395186

fbshipit-source-id: 37aaba4a46649971d8426ba20cb1b7348ddb5958