Add Rds to alias analysis
Summary:
I noticed we generate pretty bad code for static properties; both
multiple loads from the same rds location, and multiple stores to the
same location. In addition, even after emitting a CheckType<InitNull>
for a static property, we would still load the old value and decref if
necessary when writing to it.

Reviewed By: swtaarrs

Differential Revision: D8389385

fbshipit-source-id: 98159b2c12efb5b4d0369ef685808c946bfd2c90