Add call-time pass-by-reference annotations to test/run, make it hhvm-only
Summary:
The warning is now on by default, so the annotations avoid lots of
spew. And since this is a Hack-only feature, change from <?php to <?hh. Then,
since we can't have a single correct hashbang that works internally and
externally, turn test/run into a bash script that finds an hhvm binary to use
from a number of common locations.

Reviewed By: mxw, jano

Differential Revision: D8426155

fbshipit-source-id: 6a5fe169a26cbb68ed9a62bfa04f42fb805066e2