Add typehint to code property in Exception base class
Summary: We are working on consistently typing the `code` property in the `Exception` base class. Before our work on this, `code` was typed in various ways in its subclasses (as nullable, string, etc.). With the help of DavidSnider, we have made it such that `code` is typed consistently as an int or enum int in all subclasses. The final piece of this work is to type it in the base class as an int, and we should have consistent typing for this property.

Reviewed By: DavidSnider

Differential Revision: D8374360

fbshipit-source-id: 94ceef6c72909f44aaf892528b1ae21bffafa945