use pointer address to help check countedness
Summary:
Allocate persistent string/array in lowerer memory, and counted ones (including object/resource/ref) above 1TB.
This essentially allows us to encode countedness into the pointer in a TypedValue (potentially making Persistent* types redundant, if efficient code sequence is available).

This diff uses it when generating code for refcounting when a TypedValue is known to have a pointer, and profile shows that the underlying HeapObject is likely persistent.

Reviewed By: markw65

Differential Revision: D8174093

fbshipit-source-id: 020d1518c7f7c41919aa329f9c09491540f042cd