Equality check linter
Summary:
A `===` check can never return true if the intersection of the types of the LHS and RHS is empty. This linter attempts to detect whether that intersection is empty by verifying that, for a comparison between types A and B, A <: B or B <: A. When A or B are union types, we do this subtyping check for each pair of types in the cartesian product of the lists of types constituting A and B.

`C\contains`, `C\contains_key`, and `!==` checks can be verified similarly, and this linter checks them too.

Reviewed By: manzyuk

Differential Revision: D8327769

fbshipit-source-id: f31db4067cbce525ee5a8185e5f4dd02551b7dda