Produce better errors in node patterns
Summary: When parsing the input JSON, the keytrace in the children of node patterns failed to include the index of the child. This means that if there was an error parsing the JSON of a pattern deep down in the stack, you would get an inscrutable error at a field like `children.children.children`, instead of a useful field like `children.0.children.3.children`, which would tell you exactly where to look.

Reviewed By: ljw1004

Differential Revision: D8358902

fbshipit-source-id: 892e4a471f8adc8d5e586212c790b4496da47adc