Add `MissingNodePattern`
Summary:
Sometimes you want to know if a node is missing. For example, if you want to match `if`-statements without corresponding `else`-statements, you would have to assert that the `else` clause is of kind `Missing`.

`NodePattern` would probably natural work with `missing` as a kind, except that it seems to reject it because it's not actually in the schema definition. I could add a special-case to allow `missing` for `NodePattern`s, but if I have to make a change anyways, it seems better to add a dedicated `MissingNodePattern`. Then we can also forbid asserting things about the children of `Missing` (since `Missing` has no children).

Reviewed By: ljw1004

Differential Revision: D8383021

fbshipit-source-id: a2504cb795b2910e6b138f2fd4706122b78a4310