Add random_int function to HSL random extension
Summary:
This is in preparation for breaking SecureRandom's dependency on PHP. For now, we effectively reuse PHP's `random_int` implementation, which we can move into the HSL extension when we choose to remove it.

A followup diff will handle `random_bytes`, after we decide if we want to move the alphabet functionality into HHVM or keep it in Hack.

Reviewed By: markw65

Differential Revision: D8405360

fbshipit-source-id: 6fd868b5b8bf3dae33e01a84d148e90380503f2b